implement necessary code to make the web request to the necessary gist

+++ features/property_list/src/main/java/com/pedronsouza/feature/property_list/PropertyListItem.kt
@@ -3,5 +3,5 @@ package com.pedronsouza.feature.property_list
 data class PropertyListItem(
     val id: String,
     val name: String,
-    val value: Float
+    val value: Double
 )
